Changed GameObject Interface
Changed the setMVP (Model View Projection) to setVPM (View Projection
Matrix) to make it easier on the caller. The callee now handles
calculating their own MVP, or deciding not to use one. This all gives
the system for flexibility. [Also I wanted to change the stupid commit
summary I put on before]